Move files to deps package
As part of the larger Bionic file reorganization plan, this change
moves extras.py and optdep.py into a separate deps module. Bionic users
shouldn't be using these files directly, so it should be safe to move
them.
