+++ bionic/protocols.py
@@ -24,7 +24,7 @@ import pandas as pd
 
 from .decoration import decorator_updating_accumulator
 from .exception import UnsupportedSerializedValueError
-from .optdep import import_optional_dependency
+from .deps.optdep import import_optional_dependency
 from .util import (
     read_hashable_bytes_from_file_or_dir,
     oneline,
