+++ bionic/provider.py
@@ -28,7 +28,7 @@ from .descriptors.parsing import entity_dnode_from_descriptor
 from .descriptors import ast
 from .bytecode import canonical_bytecode_bytes_from_func
 from .util import groups_dict, hash_to_hex, oneline
-from .optdep import import_optional_dependency
+from .deps.optdep import import_optional_dependency
 
 import logging
